Begin implementing TTF logic; separate utilities
Add functions to read the `head` and `glyf` sections of the TTF file into
TTFReader, rather than the DefineFont3Tag class. This will separate the
logic into a more suited class, so that it can be utilized by other
classes.

Utility functions (e.g. pad, toFixed) are in a separate, global object
('Utils'), so that they can be reused throughout the application.