Fully implement single glyph reading
The readSingleGlyph function on TTFReader has been implemented, but
support for compound glyphs is still lacking.

Constants have also been added for single glyphs and flags, so that there
is no need to refer to the TTF specification or litter the code with
comments on what the flags are.

The DefineFont3 tag logic has been moved to the TTFReader (where it
belongs) and now only contains logic to write its specific fields to the
body. This, however, relies on knowing character mappings, and so the
ability to read the `cmap` table is partially implemented in the TTFReader
class.